maintenance: rename xenctrl.1.1.0 to xenctrl.master
With this change, the xen-api stockholm branch will now
compile on travis.

opam seems to prefer branch names over version
numbers, meaning that xenctrl.1.1.0 is not chosen
when you try to compile a package depending on xenctrl
- instead opam will pick xenctrl.dummy. However xenctrl
dummy does not compile, leading to failing builds.

As a work-around, we rename xenctrl.1.1.0 to xenctrl.master
because opam will pick "master" over "dummy". xenctrl is of
course still pinned for stockholm.

For clarity, master also compiles against the 'real' xenctrl.

Signed-off-by: lippirk <ben.anson@citrix.com>